+++ pasqal_cloud/__init__.py
@@ -48,12 +48,15 @@ from pasqal_cloud.job import CreateJob, Job
 from pasqal_cloud.utils.constants import JobStatus  # noqa: F401
 from pasqal_cloud.utils.filters import (
     BatchFilters,
-    CancelJobFilters,
     JobFilters,
     PaginationParams,
     RebatchFilters,
 )
-from pasqal_cloud.utils.responses import JobCancellationResponse, PaginatedResponse
+from pasqal_cloud.utils.responses import (
+    BatchCancellationResponse,
+    JobCancellationResponse,
+    PaginatedResponse,
+)
 from pasqal_cloud.workload import Workload
 
 from ._version import __version__, deprecation_date
@@ -359,6 +362,46 @@ class SDK:
             raise BatchCancellingError(e) from e
         return Batch(**batch_rsp, _client=self._client)
 
+    def cancel_batches(
+        self,
+        filters: Optional[BatchFilters] = None,
+    ) -> BatchCancellationResponse:
+        """
+        Cancel a group of batches matching the filters.
+
+        Args:
+            filters: filters to be applied to find the batches that will be cancelled
+
+        Returns:
+            BatchCancellationResponse:
+            a class containing the batches that have been cancelled and the id of
+            the batches that could not be cancelled with the reason explained
+
+        Raises:
+            BatchCancellingError which spawns from a HTTPError
+
+        """
+        if filters is None:
+            filters = BatchFilters()
+        elif not isinstance(filters, BatchFilters):
+            raise TypeError(
+                "Filters needs to be a CancelBatchFilters instance, "
+                f"not a {type(filters)}"
+            )
+
+        try:
+            response = self._client.cancel_batches(
+                filters=filters,
+            )
+        except HTTPError as e:
+            raise BatchCancellingError(e) from e
+        return BatchCancellationResponse(
+            batches=[
+                Batch(**batch, _client=self._client) for batch in response["batches"]
+            ],
+            errors=response["errors"],
+        )
+
     def rebatch(
         self,
         id: Union[UUID, str],
@@ -562,7 +605,7 @@ class SDK:
     def cancel_jobs(
         self,
         batch_id: Union[UUID, str],
-        filters: Optional[CancelJobFilters] = None,
+        filters: Optional[JobFilters] = None,
     ) -> JobCancellationResponse:
         """
         Cancel a group of jobs matching the filters in a selected batch.
@@ -581,8 +624,8 @@ class SDK:
 
         """
         if filters is None:
-            filters = CancelJobFilters()
-        elif not isinstance(filters, CancelJobFilters):
+            filters = JobFilters()
+        elif not isinstance(filters, JobFilters):
             raise TypeError(
                 "Filters needs to be a CancelJobFilters instance, "
                 f"not a {type(filters)}"